refactor: move FoundPosition class to compare.py and remove output_models.py
